Update connection-pools.md
This clears the confusion of what "each of these" refers to since there are two possible items it can apply to from the previous line. Either application modules or each front-end. Calculating one of the two items generates a completely different result so adding exactly this wording will help solve the debate.