Update how-licensing-works.md
This article is not about "licensing" in the standard sense of this term. Licensing is the contractual granting of rights to use someone else's intellectual property. What this article describes is how infrastructures and runtime environments relate and how they're identified. There are a lot of different concepts mixed together and referring to this as licensing is misleading. Happy to chat about these changes. I also have used "runtime" when referring to the environments. I find it is clearer in that it helps people understand that this is where the applications run.