Clarify deduplication
The section where we talk about deduplication (" as long as they are assigned to a default tenant, even across user providers.") is confusing to customers. A User Provider can only have 1 default tenant. Therefore, users will only be deduplicated if they belong to different User Providers (same as default tenants). With the current writing, it seems that different User Providers is an optional situation. 

From different Support cases, we can see that customers believe that this sentence actually means that users in different tenants will be counted as 1 user only (inside the same or different User Providers), as long as one of the users belongs to the Default tenant. This is not true. We only discard users as duplicates if they exist across different default tenants (which, by definition, means different user providers).